requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
more-itertools==8.13.0
multidict==6.0.2
python-decouple==3.6
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.10
yarl==1.7.2
